+++ crates/tauri-build/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## \[2.5.3]
+
+### Dependencies
+
+- Upgraded to `tauri-utils@2.8.1`
+- Upgraded to `tauri-codegen@2.5.2`
+
+## \[2.5.2]
+
+### Dependencies
+
+- Upgraded to `tauri-codegen@2.5.1`
+
 ## \[2.5.1]
 
 ### Bug Fixes